# Conflicts:
#	sshpuffin/src/ssh/message.rs
#	tlspuffin/src/protocol.rs
#	tlspuffin/src/tls/fn_messages.rs
#	tlspuffin/src/tls/mod.rs
#	tlspuffin/src/tls/seeds.rs
#	tlspuffin/src/tls/vulnerabilities.rs

+++ tlspuffin/src/tls/fn_messages.rs
@@ -191,7 +191,7 @@ nyi_fn! {
     /// hello_verify_request_RESERVED => 0x03,
 }
 /// NewSessionTicket => 0x04,
-pub fn fn_new_session_ticket(lifetime_hint: &u64, ticket: &PayloadU16) -> Result<Message, FnError> {
+pub fn fn_new_session_ticket(lifetime_hint: &u32, ticket: &PayloadU16) -> Result<Message, FnError> {
     // todo unclear where the arguments come from here, needs manual trace implementation
     //      https://github.com/tlspuffin/tlspuffin/issues/155
     Ok(Message {
@@ -199,7 +199,7 @@ pub fn fn_new_session_ticket(lifetime_hint: &u64, ticket: &PayloadU16) -> Result
         payload: MessagePayload::Handshake(HandshakeMessagePayload {
             typ: HandshakeType::NewSessionTicket,
             payload: HandshakePayload::NewSessionTicket(NewSessionTicketPayload {
-                lifetime_hint: *lifetime_hint as u32,
+                lifetime_hint: *lifetime_hint,
                 ticket: ticket.clone(),
             }),
         }),
